Move audio and sprite data from game data folder
